+++ cli/cmd/root.go
@@ -182,6 +182,7 @@ func NewCmdRoot() *cobra.Command {
 		NewSearchCmd(),
 		NewCleanCmd(),
 		NewCreateCmd(),
+		NewNewCmd(),
 		NewInstallCmd(),
 		NewUninstallCmd(),
 		NewPackageCmd(),